Single page playbook generator, simple navigation
